Fix window resizing issue and refactor handlers into a separate file
