+++ src/handlers.h
@@ -1,0 +1,9 @@
+#ifndef HANDLERS_H
+#define HANDLERS_H
+#include <SDL.h> 
+
+void handleInput(const Uint8* keyPressed);
+
+void handleResize(int w, int h);
+
+#endif
